Fixed crash in support.c:148 on 64 bit systems.
git-svn-id: https://svn.code.sf.net/p/imagination/code@147 6247ab75-7297-438c-a817-b4e4d5db8f37

+++ src/support.c
@@ -145,7 +145,7 @@ GtkWidget *_gtk_combo_box_new_text(gboolean pointer)
 	cell = gtk_cell_renderer_text_new ();
 	gtk_cell_layout_pack_start (GTK_CELL_LAYOUT (combo_box), cell, TRUE);
 	gtk_cell_layout_set_attributes (GTK_CELL_LAYOUT (combo_box), cell,"text", 0, NULL);
-	g_object_set(cell,"ypad", 0.0, NULL);
+	g_object_set(cell,"ypad", (guint)0, NULL);
 	return combo_box;
 }
